refactor(config): unify time fields as float + centralize font OCR config (#11)
* refactor(config): change time-related config fields to float

* refactor(config): move font/OCR related settings to general.font_ocr

* docs(config): update time fields to float and add general.font_ocr section

+++ novel_downloader/core/interfaces/requester_protocol.py
@@ -26,7 +26,7 @@ class RequesterProtocol(Protocol):
         """
         ...
 
-    def get_book_info(self, book_id: str, wait_time: Optional[int] = None) -> str:
+    def get_book_info(self, book_id: str, wait_time: Optional[float] = None) -> str:
         """
         Fetch the raw HTML (or JSON) of the book info page.
 
@@ -37,7 +37,7 @@ class RequesterProtocol(Protocol):
         ...
 
     def get_book_chapter(
-        self, book_id: str, chapter_id: str, wait_time: Optional[int] = None
+        self, book_id: str, chapter_id: str, wait_time: Optional[float] = None
     ) -> str:
         """
         Fetch the raw HTML (or JSON) of a single chapter.
@@ -55,7 +55,7 @@ class RequesterProtocol(Protocol):
         """
         ...
 
-    def get_bookcase(self, wait_time: Optional[int] = None) -> str:
+    def get_bookcase(self, wait_time: Optional[float] = None) -> str:
         """
         Optional: Retrieve the HTML content of the authenticated user's bookcase page.
